consolidate handling of syntax error markup
sqltoast, when it detects a syntax error in the bytestream, will
construct an error that includes the original SQL along with an
underlying string that points to the specific location of the error,
like so:

$ ./sqltoaster "CREATE DATABASE '"
Parsing failed.
In quoted identifier parsing, failed to find closing escape character
"'".
CREATE DATABASE '
                ^
Expected symbol <identifier> but found UNKNOWN

This patch segregates out the code that handles putting that marker line
together.