correct trivial segfault cause
Wasn't adding a keyword token to the tokens stack when finding a
keyword, which resulted in a segfault. Fixed that up and added in a new
PARSE_INPUT_ERROR parse_result_t value to represent issues with the
subject being parsed, the first issue of which would be an empty string.