use a deque for the tokens collection
Changes the type of the parse_context_t::tokens field to a
std::deque<token_t> to support both indexed access and reverse iteration
as well as stack-like operations to pop tokens off the top of the stack
efficiently.

First use of the stack-like behaviour is popping comment tokens off the
token stack during parse().