trim token stack when pushing statements
When successfulling parsing a statement, we dynamically allocate a
sqltoast::statement_t subclass and push it onto the
parse_result_t::statements vector. We now trim the tokens stack to reset
the front of the tokens stack to be the next non-comment token, allowing
multiple statements to be successfully parsed.