slowly encapsulating token processing
Changes the way that we test for "I'm at the end of the input stream" to
be a parse_context_t::at_end() method. This method currently checks to
see whether a supplied token iterator is pointing to the end of the
token stack. In the future, this will be replaced by a method on the
*lexer* instead of the context and will simply return whether the input
stream has been fully tokenized.

Issue #15