store an error in the lexer
I'd like to get rid of the heavyweight error checking in the current
lexer that depends on the parse_context_t struct. In order to do that, I
added a new error_t enum that will store error codes (but no error
strings).

Also removed the escape_mode member of the lexer_t since that isn't
useful outside of the token_identifier() functions.

Issue #15