No more tokens stack or tokenize()
Removes the tokens stack on the parse_context_t entirely, changing the
parsing implementations to request the next token on-demand using the
next_token() function, which returns a pointer to a token_t struct.

Still lots of work to do, including implementing a peek_symbol()
mechanism to avoid the rewinds currently in the parse implementations.

Issue #15