introduce tokenize_func_t
Introduces a function pointer type called tokenize_func_t that returns
whether a token was found of a particular category, or not found, or
whether an error in lexing occurred.

Next step is to remove the parser_context_t& parameter to the
tokenize_func_t type and replace with the lexer.

Issue #15