begin returning token_t in tokenize_result_t
The goal is to allow peek() ability on tokens or symbols, which would
enable looking for the next token or symbol *without* moving the lexer's
cursor. In order to do this, we need to begin changing the
tokenize_func_t functions to returning the found token instead of
storing found tokens in the lexer's ::current_token member.

Issue #15