make parse_references_specification() reusable
The parse_references_constraint() function wasn't very reusable because
it was relying on there being a single column definition to apply the
found references specification clause to. This meant that the parsing
code for the <references specification> clause could not be reused by
the parse_constraint() function. This patch reworks the function
signature of parse_references_constraint() so that it doesn't assume it
is being called from a single-column constraint location.