Fix NOT handling in search conditions
The handling of the reverse operation (NOT) in the various predicates
was not being done correctly. Basically, if an expression was AND'd
together and that expression had been reversed, the first expression was
reversed, not the second. Making the boolean_term_t's constructor take a
bool indicating the operation was reversed solved the problem.