Adds a simple test framework
Adds a simple test framework for SQL grammar checking. Basically, the
test runner (written in Python) reads files in tests/grammar/$DIALECT
directories that end in .test. These test files contain an input to
sqltoaster, denoted with lines that begin with the ">" character,
followed by verbatim expected output from sqltoaster that describes the
parsed statement(s).

For example, here is the initial test file for the INSERT INTO statement
that uses the DEFAULT VALUES clause:

```
>INSERT INTO t1 DEFAULT VALUES
OK
statements[0]:
  <statement: INSERT
   table name: t1
   default columns: true
   default values: true>
```

The test runner simply throws the input SQL to sqltoaster and determines
if the output matches the expected output in the test file.

Issue #1