begin breaking out parse_value_expression()
The grammar for value expressions is extremely complex. Currently, we
only parse a subsection of what's called the "value expression primary"
potential subclasses of a value expression. We want to also parse string
value expressions, datetime value expressions, interval value
expressions as well as the CASE clause and CAST specifications.

This patch simply breaks the *existing* parse_value_expression()
function out into separate smaller functions and renames the functions
to accurately represent the small part of the parse grammar the function
deals with.