support native string concatenation expressions
Adds support for the || operator to concatenate character factors into a
character value expression. This involved reworking the way that
punctuators (now called just "special symbols") were tokenized. Because
the || symbol is comprised of two vertical bar characters, I reworked
the tokenization to do a lookahead when the vertical bar is found.

Issue #66