add basics of numeric function parsing
Numeric functions are one of two types of numeric primaries. Numeric
factors, which are components of numeric expressions (they are the
operands of multi-term expressions) are composed of a sign and a numeric
primary. Numeric primaries are either a value expression primary or a
numeric function like EXTRACT, POSITION or CHAR_LENGTH.

This patch adds the parser infrastructure for parsing the length
expression functions CHAR_LENGTH and CHARACTER_LENGTH, along with tests
showing the parsed tree of numeric factors/primaries.

Issue #73