Refactors the data type descriptor parse functions
Instead of passing a reference to the containing column_definition_t
struct, we now pass a reference to a
std::unique_ptr<data_type_descriptor_t> and fill this
data_type_descriptor_t pointer with the subclasses of data types.

Next step is to remove the usage of passing vectors of constraints and
column definitions around.