Move printer functions into sqltoaster
Moves the myriad printer functions (operator<<() overloads for various
structs to std::ostream) into the sqltoaster binary. The sqltoaster
binary is the thing that should dictate the printing format for the
sqltoast objects, not the library itself, which should focus strictly on
the structure of the objects and the parsing functionality.