NOT symbol is a value expression terminator
The NOT keyword is a terminator for value expressions. This was missing
from the value expression parsers, which was causing the following to
fail parsing:

 SELECT * FROM t1 WHERE a NOT IN (1)

but this would pass:

 SELECT * FROM t1 WHERE NOT a IN (1)

The reason is because "a" is a value expression (a column reference) and
the NOT symbol was being considered part of that value expression
instead of an indication that the value expression should be terminated.

Fixes Issue #94