prep for row value constructor list and subquery
A row value constructor may actually be one of three things. The first,
and most common, is a row value constructor element, which is
essentially a value expression that reduces to a scalar value. The
second is a list of row value constructor elements. The third is a
scalar subquery.

This patch adds accomodation for this by making row_value_constructor
into row_value_constructor_element and adding another layer of
type-linkage for row value constructor list and row subquery types that
will come in followup patches.

Issue #96
Issue #97