Work list item comprehension into YAML printer
The idea behind this patch is to lay the groundwork for a more generic
approach to outputting the YAML for list or object attributes. As we get
more and more YAML printers added for the finest-grained query and value
expression components, we start to need a better way of the to_yaml()
functions being able to recognize whether the element they are
outputting in YAML is a list item or not.

The first and most obvious query component to change is the top-level
statement_t to_yaml() printer.

Issue #104