Start refactoring the YAML printer
I was never really happy with the sqltoaster::printer_t struct and the
use of iomanips to control indentation levels and list construction. So,
I'm going to rewrite the YAML printing to instead use a stack push/pop
system of generic YAML node elements.

The first step to this is to get rid of storing the output format in an
iomanip xalloc bucket, which was just sill to begin with.

Issue #104