correct INSERT statement parsing
I made a silly mistake in the INSERT statement parser resulting from a
misreading of the SQL-92 EBNF grammar for the INSERT statement. There is
no such thing as an INSERT_SELECT statement type. An INSERT statement
always has a <query expression> component. It's just that with the
INSERT ..  VALUES (...) style statement, the VALUES (...) clause *is* a
<query expression>. The VALUE (...) clause is called a <table value
constructor> and is a type of <non join query primary>.

This patch reduces the parsing code in parse_insert() and breaks out the
parsing of table value constructors into a separate function.

In modifying the test files, I discovered a few bugs with how character
and datetime expressions as well as numeric expressions involving
parentheses are processed. These expressions are not properly converted
into <row value constructor element> components due to the incorrect
recognition of the closing RPAREN. I'll fix this in a followup patch.