Start of new YAML printer strategy
Instead of using a printer_t to maintain state on indentation and list
levels in iomanip xallocs, I want to move to an easier-to-understand
YAML printing strategy. This patch adds three structs representing types
of nodes that are produced in YAML documents: a scalar value, a mapping
node, and a sequence node.

The patch includes rewrites of all the (still existing) to_yaml() YAML
printer functions to instead add sequence, scalar, and mapping subnodes
to a supplied mapping node representing the attributes for a particular
parse tree element.

Issue #104