Break out query_term_t and union/except parsers
Most of the patch is test file updates after doing the necessary
breakout of the YAML printers for query_term_t and the various
query_expression_t subtypes.

We begin the process of parsing UNION and EXCEPT query expressions but
as of now do not include functional tests for those expression subtypes.
That will come in a following patch, since this is already getting a
little long in the tooth.

Issue #78