add cmake foo for libunwind
Despite improvements in cmake, it's still annoyingly difficult to use
the FIND_PACKAGE and FIND_LIBRARY macros. One annoyance is that
FIND_LIBRARY and FIND_PACKAGE don't work per-build-type, which means you
need to keep debug-specific linking within if() blocks.

I was able to find a custom Findlibunwind.cmake find module online and
have imported that into the project, and modified the various
CMakeLists.txt files accordingly to only link libunwind when we're in a
debug build.

Issue #129